ImageRetrievalSystem/requirements.txt
clip==1.0
Pillow==9.4.0
Pillow==9.5.0
prettytable==3.7.0
torch==2.0.1
